logproto: add logproto-framed-proxied-text-server
