+++ lib/logproto/logproto-builtins.c
@@ -25,6 +25,7 @@
 #include "logproto-text-client.h"
 #include "logproto-text-server.h"
 #include "logproto-proxied-text-server.h"
+#include "logproto-framed-proxied-text-server.h"
 #include "logproto-framed-client.h"
 #include "logproto-framed-server.h"
 #include "plugin.h"
@@ -40,6 +41,7 @@ DEFINE_LOG_PROTO_SERVER(log_proto_text);
 DEFINE_LOG_PROTO_SERVER(log_proto_text_with_nuls);
 DEFINE_LOG_PROTO_SERVER(log_proto_proxied_text);
 DEFINE_LOG_PROTO_SERVER(log_proto_proxied_text_tls_passthrough, .use_multitransport = TRUE);
+DEFINE_LOG_PROTO_SERVER(log_proto_framed_proxied_text);
 DEFINE_LOG_PROTO_CLIENT(log_proto_framed);
 DEFINE_LOG_PROTO_SERVER(log_proto_framed);
 
@@ -55,6 +57,7 @@ static Plugin framed_server_plugins[] =
   LOG_PROTO_SERVER_PLUGIN(log_proto_proxied_text_tls_passthrough, "proxied-tls-passthrough"),
   LOG_PROTO_CLIENT_PLUGIN(log_proto_framed, "framed"),
   LOG_PROTO_SERVER_PLUGIN(log_proto_framed, "framed"),
+  LOG_PROTO_SERVER_PLUGIN(log_proto_framed_proxied_text, "proxied-framed"),
 };
 
 void
