+++ lib/logproto/logproto-proxied-text-server.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2020 One Identity
+ * Copyright (c) 2020-2023 One Identity LLC.
  * Copyright (c) 2022 Balazs Scheidler <bazsi77@gmail.com>
  *
  * This library is free software; you can redistribute it and/or
@@ -22,101 +22,7 @@
  *
  */
 
-
-#include <stdio.h>
-#include <stdlib.h>
-#include "messages.h"
-#include "logproto-proxied-text-server.h"
-#include "transport/multitransport.h"
-#include "transport/transport-factory-tls.h"
-#include "str-utils.h"
-
-#define IP_BUF_SIZE 64
-
-/* This file implements: https://www.haproxy.org/download/1.8/doc/proxy-protocol.txt */
-
-/* PROXYv1 line without newlines or terminating zero character.  The
- * protocol specification contains the number 108 that includes both the
- * CRLF sequence and the NUL */
-#define PROXY_PROTO_HDR_MAX_LEN_RFC 105
-
-/* the size of the buffer we use to fetch the PROXY header into */
-#define PROXY_PROTO_HDR_BUFFER_SIZE 1500
-
-/* the amount of bytes we need from the client to detect protocol version */
-#define PROXY_PROTO_HDR_MAGIC_LEN   5
-
-typedef struct _LogProtoProxiedTextServer
-{
-  LogProtoTextServer super;
-
-  // Info received from the proxied that should be added as LogTransportAuxData to
-  // any message received through this server instance.
-  struct
-  {
-    gboolean unknown;
-
-    gchar src_ip[IP_BUF_SIZE];
-    gchar dst_ip[IP_BUF_SIZE];
-
-    int ip_version;
-    int src_port;
-    int dst_port;
-  } info;
-
-  // Flag to only run handshake() once
-  gboolean handshake_done;
-  gboolean has_to_switch_to_tls;
-
-  enum
-  {
-    LPPTS_INITIAL,
-    LPPTS_DETERMINE_VERSION,
-    LPPTS_PROXY_V1_READ_LINE,
-    LPPTS_PROXY_V2_READ_HEADER,
-    LPPTS_PROXY_V2_READ_PAYLOAD,
-  } header_fetch_state;
-
-  /* 0 unknown, 1 or 2 indicate proxy header version */
-  gint proxy_header_version;
-  guchar proxy_header_buff[PROXY_PROTO_HDR_BUFFER_SIZE];
-  gsize proxy_header_buff_len;
-} LogProtoProxiedTextServer;
-
-
-struct proxy_hdr_v2
-{
-  uint8_t sig[12];  /* hex 0D 0A 0D 0A 00 0D 0A 51 55 49 54 0A */
-  uint8_t ver_cmd;  /* protocol version and command */
-  uint8_t fam;      /* protocol family and address */
-  uint16_t len;     /* number of following bytes part of the header */
-};
-
-union proxy_addr
-{
-  struct
-  {
-    /* for TCP/UDP over IPv4, len = 12 */
-    uint32_t src_addr;
-    uint32_t dst_addr;
-    uint16_t src_port;
-    uint16_t dst_port;
-  } ipv4_addr;
-  struct
-  {
-    /* for TCP/UDP over IPv6, len = 36 */
-    uint8_t  src_addr[16];
-    uint8_t  dst_addr[16];
-    uint16_t src_port;
-    uint16_t dst_port;
-  } ipv6_addr;
-  struct
-  {
-    /* for AF_UNIX sockets, len = 216 */
-    uint8_t src_addr[108];
-    uint8_t dst_addr[108];
-  } unix_addr;
-};
+#include "logproto-proxied-text-server-private.h"
 
 #define PROXY_HDR_TCP4 "PROXY TCP4 "
 #define PROXY_HDR_TCP6 "PROXY TCP6 "
@@ -588,7 +494,7 @@ _augment_aux_data(LogProtoProxiedTextServer *self, LogTransportAuxData *aux)
   return;
 }
 
-static LogProtoStatus
+LogProtoStatus
 log_proto_proxied_text_server_fetch(LogProtoServer *s, const guchar **msg, gsize *msg_len, gboolean *may_read,
                                     LogTransportAuxData *aux, Bookmark *bookmark)
 {
@@ -605,7 +511,7 @@ log_proto_proxied_text_server_fetch(LogProtoServer *s, const guchar **msg, gsize
   return LPS_SUCCESS;
 }
 
-static void
+void
 log_proto_proxied_text_server_free(LogProtoServer *s)
 {
   LogProtoProxiedTextServer *self = (LogProtoProxiedTextServer *) s;
@@ -617,7 +523,7 @@ log_proto_proxied_text_server_free(LogProtoServer *s)
   return;
 }
 
-static void
+void
 log_proto_proxied_text_server_init(LogProtoProxiedTextServer *self, LogTransport *transport,
                                    const LogProtoServerOptions *options)
 {
